UI: Data Table, move KeyboardNav into Data.js
