+++ src/UI/templates/js/Table/src/table.js
@@ -2,7 +2,6 @@ import il from 'il';
 import $ from 'jquery';
 import Params from './Params';
 import Data from './Data';
-import Keyboardnav from './Keyboardnav';
 
 il.UI = il.UI || {};
 il.UI.table = il.UI.table || {};
@@ -10,5 +9,4 @@ il.UI.table = il.UI.table || {};
 il.UI.table.data = new Data(
   $,
   new Params(),
-  new Keyboardnav(),
 );
